+++ src/ff.h
@@ -25,7 +25,7 @@
  * @param pending_process_queue
  * @param process_queue
  */
-void fc_fs(deque_t *pending_process_queue, deque_t *process_queue, char *memory_opt, int memory_size);
+void fc_fs(deque_t *pending_process_queue, deque_t *process_queue, char *memory_opt, long long memory_size);
 
 /**
  * Decrements a processes time remaining attribute
@@ -45,8 +45,8 @@ void run_process_ff(process_t *process);
  * @param next_process_arrival_time
  * @return
  */
-void step_ff(deque_t *process_queue, process_t *current_process, int *simulation_time_elapsed, int *pages,
-        int num_pages, int *space_available, int *state, int *loading_cost, char *memory_opt, int *pages_time, int *loading_status);
+void step_ff(deque_t *process_queue, process_t *current_process, long long *simulation_time_elapsed, long long *pages,
+        long long num_pages, long long *space_available, long long *state, long long *loading_cost, char *memory_opt, long long *pages_time, long long *loading_status);
 
 /**
  * Takes processes from pending queue and adds them to the process queue to simulate the arrival of a new process
@@ -54,7 +54,7 @@ void step_ff(deque_t *process_queue, process_t *current_process, int *simulation
  * @param process_queue
  * @return
  */
-void check_pending(deque_t *pending_process_queue, deque_t *process_queue, int simulation_time);
+void check_pending(deque_t *pending_process_queue, deque_t *process_queue, long long simulation_time);
 
 
 #endif
